requirements.txt
# Medical KAG System Dependencies
# Version: 7.13 (Spine GraphRAG - Single Store: Neo4j + DOI Fulltext)
# Last Updated: 2025-12-25

# ============================================
# LLM Integration
# ============================================
anthropic>=0.40.0                 # Claude Haiku 4.5 / Sonnet (Primary)
openai>=1.0.0                     # OpenAI Embeddings (text-embedding-3-large)
python-dotenv>=1.0.0
pydantic>=2.0

# ============================================
# Graph Database (Neo4j - Single Store)
# ============================================
neo4j>=5.15.0                     # Neo4j Python Driver (async support)
nest-asyncio>=1.6.0               # Async event loop nesting for Neo4j vector search

# ============================================
# PDF Processing
# ============================================
PyMuPDF>=1.23.0

# ============================================
# Web / HTTP
# ============================================
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
aiohttp>=3.9.0

# ============================================
# Embedding & ML
# ============================================
sentence-transformers>=2.2.0      # Local medical embeddings (MedTE)
torch>=2.0.0

# ============================================
# MCP Server (SSE Transport)
# ============================================
mcp>=1.0.0
starlette>=0.27.0                 # ASGI framework for SSE
uvicorn>=0.23.0                   # ASGI server

# ============================================
# Utilities
# ============================================
pyyaml>=6.0.0
python-dateutil>=2.8.0
jsonschema>=4.0.0
rapidfuzz>=3.0.0                  # Fuzzy string matching for entity normalization

# ============================================
# Testing
# ============================================
pytest>=7.0.0
pytest-asyncio>=0.21.0

# ============================================
# Web UI (Streamlit)
# ============================================
streamlit>=1.28.0
plotly>=5.18.0
networkx>=3.0
